Display permissions tooltip on Add Integration button in Agent Policy view
