+++ x-pack/plugins/fleet/public/applications/fleet/sections/agent_policy/details_page/components/package_policies/package_policies_table.tsx
@@ -24,8 +24,18 @@ import {
 import { INTEGRATIONS_PLUGIN_ID } from '../../../../../../../../common';
 import { pagePathGetters } from '../../../../../../../constants';
 import type { AgentPolicy, InMemoryPackagePolicy, PackagePolicy } from '../../../../../types';
-import { PackageIcon, PackagePolicyActionsMenu } from '../../../../../components';
-import { useAuthz, useLink, usePackageInstallations, useStartServices } from '../../../../../hooks';
+import {
+  EuiButtonWithTooltip,
+  PackageIcon,
+  PackagePolicyActionsMenu,
+} from '../../../../../components';
+import {
+  useAuthz,
+  useLink,
+  usePackageInstallations,
+  usePermissionCheck,
+  useStartServices,
+} from '../../../../../hooks';
 import { pkgKeyFromPackageInfo } from '../../../../../services';
 
 interface Props {
@@ -54,6 +64,10 @@ export const PackagePoliciesTable: React.FunctionComponent<Props> = ({
   const canReadIntegrationPolicies = useAuthz().integrations.readIntegrationPolicies;
   const { updatableIntegrations } = usePackageInstallations();
   const { getHref } = useLink();
+
+  const permissionCheck = usePermissionCheck();
+  const missingSecurityConfiguration =
+    !permissionCheck.data?.success && permissionCheck.data?.error === 'MISSING_SECURITY';
 
   // With the package policies provided on input, generate the list of package policies
   // used in the InMemoryTable (flattens some values for search) as well as
@@ -265,7 +279,7 @@ export const PackagePoliciesTable: React.FunctionComponent<Props> = ({
         toolsRight: agentPolicy.is_managed
           ? []
           : [
-              <EuiButton
+              <EuiButtonWithTooltip
                 key="addPackagePolicyButton"
                 fill
                 isDisabled={!canWriteIntegrationPolicies}
@@ -277,12 +291,29 @@ export const PackagePoliciesTable: React.FunctionComponent<Props> = ({
                   });
                 }}
                 data-test-subj="addPackagePolicyButton"
+                tooltip={
+                  !canWriteIntegrationPolicies
+                    ? {
+                        content: missingSecurityConfiguration ? (
+                          <FormattedMessage
+                            id="xpack.fleet.epm.addPackagePolicyButtonSecurityRequiredTooltip"
+                            defaultMessage="To add Elastic Agent Integrations, you must have security enabled and have the All privilege for Fleet. Contact your administrator."
+                          />
+                        ) : (
+                          <FormattedMessage
+                            id="xpack.fleet.epm.addPackagePolicyButtonPrivilegesRequiredTooltip"
+                            defaultMessage="Elastic Agent Integrations require the All privilege for Fleet and All privilege for Integrations. Contact your administrator."
+                          />
+                        ),
+                      }
+                    : undefined
+                }
               >
                 <FormattedMessage
                   id="xpack.fleet.policyDetails.addPackagePolicyButtonText"
                   defaultMessage="Add integration"
                 />
-              </EuiButton>,
+              </EuiButtonWithTooltip>,
             ],
         box: {
           incremental: true,
